+++ src/Buckle/Buckle.Tests/CodeAnalysis/Evaluating/EvaluatorTests.cs
@@ -1,23 +1,9 @@
-using System;
-using System.Collections.Generic;
-using Buckle.CodeAnalysis;
-using Buckle.CodeAnalysis.Evaluating;
-using Buckle.CodeAnalysis.Symbols;
-using Buckle.CodeAnalysis.Syntax;
-using Buckle.Diagnostics;
-using Diagnostics;
 using Xunit;
-using Xunit.Abstractions;
+using static Buckle.Tests.Assertions;
 
-namespace Buckle.Tests.CodeAnalysis;
+namespace Buckle.Tests.CodeAnalysis.Evaluating;
 
-public sealed partial class EvaluatorTests {
-    private readonly ITestOutputHelper writer;
-
-    public EvaluatorTests(ITestOutputHelper writer) {
-        this.writer = writer;
-    }
-
+public sealed class EvaluatorTests {
     [Theory]
     // Empty expressions
     [InlineData(";", null)]
@@ -329,89 +315,5 @@ public sealed partial class EvaluatorTests {
     [InlineData("var cond = true; int res = 3; while (true) { if (cond) break; else continue; res = 4; } return res;", 3)]
     public void Evaluator_Computes_CorrectValues(string text, object expectedValue) {
         AssertValue(text, expectedValue);
-    }
-
-    private void AssertValue(string text, object expectedValue) {
-        var syntaxTree = SyntaxTree.Parse(text);
-        var compilation = Compilation.CreateScript(null, syntaxTree);
-        var variables = new Dictionary<VariableSymbol, EvaluatorObject>();
-        var _ = false;
-        var result = compilation.Evaluate(variables, ref _);
-
-        if (result.value is double && (Convert.ToDouble(expectedValue)).CompareTo(result.value) == 0)
-            expectedValue = Convert.ToDouble(expectedValue);
-
-        Assert.Empty(result.diagnostics.FilterOut(DiagnosticType.Warning).ToArray());
-        Assert.Equal(expectedValue, result.value);
-    }
-
-    private void AssertExceptions(string text, params Exception[] exceptions) {
-        var syntaxTree = SyntaxTree.Parse(text);
-        var compilation = Compilation.CreateScript(null, syntaxTree);
-        var _ = false;
-        var result = compilation.Evaluate(new Dictionary<VariableSymbol, EvaluatorObject>(), ref _);
-
-        if (exceptions.Length != result.exceptions.Count) {
-            writer.WriteLine($"Input: {text}");
-
-            foreach (var exception in result.exceptions)
-                writer.WriteLine($"Exception ({exception}): {exception.Message}");
-        }
-
-        Assert.Equal(exceptions.Length, result.exceptions.Count);
-
-        for (int i=0; i<exceptions.Length; i++)
-            Assert.Equal(exceptions[i].GetType(), result.exceptions[i].GetType());
-    }
-
-    private void AssertDiagnostics(string text, string diagnosticText, bool assertWarnings = false) {
-        var annotatedText = AnnotatedText.Parse(text);
-        var syntaxTree = SyntaxTree.Parse(annotatedText.text);
-
-        var tempDiagnostics = new BelteDiagnosticQueue();
-
-        if (syntaxTree.diagnostics.FilterOut(DiagnosticType.Warning).Any()) {
-            tempDiagnostics.Move(syntaxTree.diagnostics);
-        } else {
-            var compilation = Compilation.CreateScript(null, syntaxTree);
-            var _ = false;
-            var result = compilation.Evaluate(new Dictionary<VariableSymbol, EvaluatorObject>(), ref _);
-            tempDiagnostics = result.diagnostics;
-        }
-
-        var expectedDiagnostics = AnnotatedText.UnindentLines(diagnosticText);
-
-        if (annotatedText.spans.Length != expectedDiagnostics.Length)
-            throw new Exception("must mark as many spans as there are diagnostics");
-
-        var diagnostics = assertWarnings
-            ? tempDiagnostics
-            : tempDiagnostics.FilterOut(DiagnosticType.Warning);
-
-        if (expectedDiagnostics.Length != diagnostics.count) {
-            writer.WriteLine($"Input: {annotatedText.text}");
-
-            foreach (var diagnostic in diagnostics.AsList())
-                writer.WriteLine($"Diagnostic ({diagnostic.info.severity}): {diagnostic.message}");
-        }
-
-        Assert.Equal(expectedDiagnostics.Length, diagnostics.count);
-
-        for (int i=0; i<expectedDiagnostics.Length; i++) {
-            var diagnostic = diagnostics.Pop();
-
-            var expectedMessage = expectedDiagnostics[i];
-            var actualMessage = diagnostic.message;
-            Assert.Equal(expectedMessage, actualMessage);
-
-            var expectedSpan = annotatedText.spans[i];
-            var actualSpan = diagnostic.location.span;
-            writer.WriteLine($"start: {expectedSpan.start}, {actualSpan.start}");
-            Assert.Equal(expectedSpan.start, actualSpan.start);
-            writer.WriteLine($"end: {expectedSpan.end}, {actualSpan.end}");
-            Assert.Equal(expectedSpan.end, actualSpan.end);
-            writer.WriteLine($"length: {expectedSpan.length}, {actualSpan.length}");
-            Assert.Equal(expectedSpan.length, actualSpan.length);
-        }
     }
 }
